Move functions to fed_utils.py
